Fix: bill test failing

+++ src/main/java/example/medCashFlow/dto/bill/BillRegisterDTO.java
@@ -1,11 +1,10 @@
 package example.medCashFlow.dto.bill;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record BillRegisterDTO(
         String name,
-        BigDecimal pricing,
+        Double pricing,
         String type,
         Long involvedId,
         Long accountPlanningId,
