+++ src/main/java/example/medCashFlow/dto/bill/BillResponseDTO.java
@@ -1,12 +1,11 @@
 package example.medCashFlow.dto.bill;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record BillResponseDTO(
         Long id,
         String name,
-        BigDecimal pricing,
+        Double pricing,
         String type,
         Long employeeId,
         Long involvedId,
